Add PredefinedBuilders property populate it with predefined builders
Currently we put the predefined builders to the Customization
scope instead of Engine because we don't want AutoProperties command
to be applied to those instances.

That leads to the necessity to always use
fixture.Customizations.Insert(0, builder) to add an own
customization to ensure we put it before the default builders.

In this commit we introduce a new collection and extract all the
predefined builders to there. Now you are free to use
fixture.Customizations.Add(builder) and you will surely run before
the predefined builders.